backend/pyproject.toml
[project]
name = "device-reservation"
version = "0.1.0"
description = "室内装置予約システム PoC"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "pydantic[email]>=2.10.0",
    "pydantic-settings>=2.6.0",
    "sqlalchemy[asyncio]>=2.0.0",
    "asyncpg>=0.30.0",
    "python-jose[cryptography]>=3.3.0",
    "httpx>=0.27.0",
]

[project.optional-dependencies]
test = [
    "ruff>=0.9.0",
    "pytest>=8.3.0",
    "pytest-cov>=6.0.0",
    "pytest-asyncio>=0.25.0",
    "httpx>=0.27.0",
]

[tool.ruff]
target-version = "py313"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "W", "I", "N", "UP", "B", "C4", "A"]
ignore = []

[tool.ruff.lint.per-file-ignores]
"src/app/routers/*.py" = ["B008", "B904"]
"src/app/auth.py" = ["B008", "B904"]
"tests/test_devices.py" = ["B017"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
pythonpath = ["src"]
testpaths = ["tests"]
filterwarnings = [
    "ignore::DeprecationWarning",
]

[tool.coverage.run]
source = ["src"]
omit = ["tests/*"]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise AssertionError",
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
    "if TYPE_CHECKING:",
]

[tool.ty]

[dependency-groups]
dev = [
    "ty>=0.0.29",
]
# ty is used for strict static type checking with pydantic
# Run: ty check src/